Add taskHelper.js to hold getTasksForTeams and getTasksForSingleUser functions
