+++ src/controllers/taskController.js
@@ -1,8 +1,7 @@
 const mongoose = require('mongoose');
-const moment = require('moment-timezone');
-const myteam = require('../helpers/helperModels/myTeam');
 const wbs = require('../models/wbs');
 const timeEntryHelper = require('../helpers/timeEntryHelper')();
+const taskHelper = require('../helpers/taskHelper')();
 const hasPermission = require('../utilities/permissions');
 
 const taskController = function (Task) {
@@ -1066,296 +1065,14 @@ const taskController = function (Task) {
 
   const getTasksForTeamsByUser = async (req, res) => {
     try {
-      const pdtstart = moment()
-        .tz('America/Los_Angeles')
-        .startOf('week')
-        .format('YYYY-MM-DD');
-      const pdtend = moment()
-        .tz('America/Los_Angeles')
-        .endOf('week')
-        .format('YYYY-MM-DD');
       const userId = mongoose.Types.ObjectId(req.params.userId);
-      const agg = myteam.aggregate([
-        {
-          $match: {
-            _id: userId,
-          },
-        },
-        {
-          $unwind: '$myteam',
-        },
-        {
-          $project: {
-            _id: 0,
-            personId: '$myteam._id',
-            name: '$myteam.fullName',
-            role: '$role',
-          },
-        },
-        // have personId, name, role
-        {
-          $lookup: {
-            from: 'userProfiles',
-            localField: 'personId',
-            foreignField: '_id',
-            as: 'persondata',
-          },
-        },
-        {
-          $project: {
-            personId: 1,
-            name: 1,
-            role: 1,
-            weeklycommittedHours: {
-              $sum: [
-                {
-                  $arrayElemAt: ['$persondata.weeklycommittedHours', 0],
-                },
-                {
-                  $ifNull: [{ $arrayElemAt: ['$persondata.missedHours', 0] }, 0],
-                },
-              ],
-            },
-          },
-        },
-        {
-          $lookup: {
-            from: 'timeEntries',
-            localField: 'personId',
-            foreignField: 'personId',
-            as: 'timeEntryData',
-          },
-        },
-        {
-          $project: {
-            personId: 1,
-            name: 1,
-            weeklycommittedHours: 1,
-            role: 1,
-            timeEntryData: {
-              $filter: {
-                input: '$timeEntryData',
-                as: 'timeentry',
-                cond: {
-                  $and: [
-                    {
-                      $gte: ['$$timeentry.dateOfWork', pdtstart],
-                    },
-                    {
-                      $lte: ['$$timeentry.dateOfWork', pdtend],
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        },
-        {
-          $unwind: {
-            path: '$timeEntryData',
-            preserveNullAndEmptyArrays: true,
-          },
-        },
-        {
-          $project: {
-            personId: 1,
-            name: 1,
-            weeklycommittedHours: 1,
-            role: 1,
-            totalSeconds: {
-              $cond: [
-                {
-                  $gte: ['$timeEntryData.totalSeconds', 0],
-                },
-                '$timeEntryData.totalSeconds',
-                0,
-              ],
-            },
-            isTangible: {
-              $cond: [
-                {
-                  $gte: ['$timeEntryData.totalSeconds', 0],
-                },
-                '$timeEntryData.isTangible',
-                false,
-              ],
-            },
-          },
-        },
-        {
-          $addFields: {
-            tangibletime: {
-              $cond: [
-                {
-                  $eq: ['$isTangible', true],
-                },
-                '$totalSeconds',
-                0,
-              ],
-            },
-            intangibletime: {
-              $cond: [
-                {
-                  $eq: ['$isTangible', false],
-                },
-                '$totalSeconds',
-                0,
-              ],
-            },
-          },
-        },
-        {
-          $group: {
-            _id: {
-              personId: '$personId',
-              weeklycommittedHours: '$weeklycommittedHours',
-              name: '$name',
-              role: '$role',
-            },
-            totalSeconds: {
-              $sum: '$totalSeconds',
-            },
-            tangibletime: {
-              $sum: '$tangibletime',
-            },
-            intangibletime: {
-              $sum: '$intangibletime',
-            },
-          },
-        },
-        {
-          $project: {
-            _id: 0,
-            personId: '$_id.personId',
-            name: '$_id.name',
-            weeklycommittedHours: '$_id.weeklycommittedHours',
-            role: '$_id.role',
-            totaltime_hrs: {
-              $divide: ['$totalSeconds', 3600],
-            },
-            totaltangibletime_hrs: {
-              $divide: ['$tangibletime', 3600],
-            },
-            totalintangibletime_hrs: {
-              $divide: ['$intangibletime', 3600],
-            },
-            percentagespentintangible: {
-              $cond: [
-                {
-                  $eq: ['$totalSeconds', 0],
-                },
-                0,
-                {
-                  $multiply: [
-                    {
-                      $divide: ['$tangibletime', '$totalSeconds'],
-                    },
-                    100,
-                  ],
-                },
-              ],
-            },
-          },
-        },
-        {
-          $lookup: {
-            from: 'tasks',
-            localField: 'personId',
-            foreignField: 'resources.userID',
-            as: 'tasks',
-          },
-        },
-        {
-          $project: {
-            tasks: {
-              resources: {
-                profilePic: 0,
-              },
-            },
-          },
-        },
-        {
-          $unwind: {
-            path: '$tasks',
-            preserveNullAndEmptyArrays: true,
-          },
-        },
-        {
-          $lookup: {
-            from: 'wbs',
-            localField: 'tasks.wbsId',
-            foreignField: '_id',
-            as: 'projectId',
-          },
-        },
-        {
-          $addFields: {
-            'tasks.projectId': {
-              $cond: [
-                { $ne: ['$projectId', []] },
-                { $arrayElemAt: ['$projectId', 0] },
-                '$tasks.projectId',
-              ],
-            },
-          },
-        },
-        {
-          $project: {
-            projectId: 0,
-            tasks: {
-              projectId: {
-                _id: 0,
-                isActive: 0,
-                modifiedDatetime: 0,
-                wbsName: 0,
-                createdDatetime: 0,
-                __v: 0,
-              },
-            },
-          },
-        },
-        {
-          $addFields: {
-            'tasks.projectId': '$tasks.projectId.projectId',
-          },
-        },
-        {
-          $lookup: {
-            from: 'taskNotifications',
-            localField: 'tasks._id',
-            foreignField: 'taskId',
-            as: 'tasks.taskNotifications',
-          },
-        },
-
-        {
-          $group: {
-            _id: '$personId',
-            tasks: { $push: '$tasks' },
-            data: {
-              $first: '$$ROOT',
-            },
-          },
-        },
-        {
-          $addFields: {
-            'data.tasks': {
-              $filter: {
-                input: '$tasks',
-                as: 'task',
-                cond: { $ne: ['$$task', {}] },
-              },
-            },
-          },
-        },
-        {
-          $replaceRoot: {
-            newRoot: '$data',
-          },
-        },
-      ]);
-      const data = await agg.exec();
-      res.status(200).send(data);
+      const teamsData = await taskHelper.getTasksForTeams(userId).exec();
+      if (teamsData.length > 0) {
+        res.status(200).send(teamsData);
+      } else {
+        const singleUserData = await taskHelper.getTasksForSingleUser(userId).exec();
+        res.status(200).send(singleUserData);
+      }
     } catch (error) {
       res.status(400).send(error);
     }
